Extract config from bump-dependencies :wrench:
